feat(g2 4.0): remove interface into base
